# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Cron Metrics Collector & Exporter API
  description: |
    A Go-based API and web server to centralize cron job results and export their statuses as Prometheus-compatible metrics.

    ## Authentication

    The API uses a two-tier authentication system:

    ### Admin API Keys
    - **Purpose**: Job management operations (CRUD)
    - **Header**: `Authorization: Bearer <admin-api-key>`
    - **Endpoints**: All `/api/job*` endpoints

    ### Per-Job API Keys
    - **Purpose**: Job result submissions (isolated per job)
    - **Header**: `X-API-Key <job-specific-api-key>`
    - **Endpoints**: `/api/job-result`
    - **Security**: Jobs can only submit results for themselves

  version: 0.3.0
  contact:
    name: Cron Exporter API Support
    url: https://github.com/jaepetto/cron-exporter
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://cronmetrics.example.com
    description: Production server
paths:
...
  /api/job:
    get:
      summary: List all jobs
      description: Retrieve a list of all registered cron jobs with optional label filtering
      tags:
        - Job Management
      security:
        - AdminAPIKey: []
      parameters:
        - name: label
          in: query
          description: "Filter jobs by label (format: key=value)"
          required: false
          schema:
            type: string
            example: "env=prod"
      responses:
        '200':
          description: Successfully retrieved job list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Job'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create a new job
      description: Register a new cron job with automatic API key generation
      tags:
        - Job Management
      security:
        - AdminAPIKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateJobRequest'
      responses:
        '201':
          description: Job created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/job-result:
    post:
      summary: Submit job execution result
      description: Submit the result of a cron job execution using the job's unique API key
      tags:
        - Job Results
      security:
        - JobAPIKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobResult'
      responses:
        '201':
          description: Job result submitted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # Public Endpoints (No Authentication Required)
...
components:
  securitySchemes:
    AdminAPIKey:
      type: http
      scheme: bearer
      description: Admin API key for job management operations

    JobAPIKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: Per-job API key for result submissions

    DashboardBasic:
      type: http
      scheme: basic
      description: Dashboard username and an admin API key as the password
  schemas:
    Job:
      type: object
      properties:
        id:
          type: integer
          description: Auto-incrementing job ID
          example: 1
        job_name:
          type: string
          description: Unique job name
          example: "daily-backup"
        host:
          type: string
          description: Host where the job runs
          example: "db-server-01"
        api_key:
          type: string
          description: Per-job API key for result submissions
          example: "cm_abc123456789abcdef123456789abcdef123456789abcd"
        automatic_failure_threshold:
          type: integer
          description: Seconds after which job is considered failed if no result submitted
          example: 3600
        labels:
          type: object
          additionalProperties:
            type: string
          description: User-defined labels for the job
          example:
            env: "prod"
            team: "platform"
            type: "backup"
        status:
          type: string
          enum: ["active", "maintenance", "paused"]
          description: Job lifecycle status
          example: "active"
        last_reported_at:
          type: string
          format: date-time
          description: Timestamp of last result submission
          example: "2025-10-30T19:56:00Z"
        created_at:
          type: string
          format: date-time
          description: Job creation timestamp
          example: "2025-10-30T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: Job last update timestamp
          example: "2025-10-30T19:56:00Z"
      required:
        - id
        - job_name
        - host
        - automatic_failure_threshold
        - labels
        - status
        - last_reported_at
        - created_at
        - updated_at
    CreateJobRequest:
      type: object
      properties:
        job_name:
          type: string
          description: Unique job name
          example: "daily-backup"
        host:
          type: string
          description: Host where the job runs
          example: "db-server-01"
        api_key:
          type: string
          description: Custom API key (optional, auto-generated if not provided)
          example: "cm_custom123456789abcdef123456789abcdef123456789abc"
        automatic_failure_threshold:
          type: integer
          description: "Seconds after which job is considered failed (default: 3600)"
          example: 7200
        labels:
          type: object
          additionalProperties:
            type: string
          description: User-defined labels for the job
          example:
            env: "prod"
            team: "platform"
            type: "backup"
        status:
          type: string
          enum: ["active", "maintenance", "paused"]
          description: "Job lifecycle status (default: active)"
          example: "active"
      required:
        - job_name
        - host
...
    JobResult:
      type: object
      properties:
        job_name:
          type: string
          description: Name of the job (must match API key)
          example: "daily-backup"
        host:
          type: string
          description: Host name (must match API key)
          example: "db-server-01"
        status:
          type: string
          enum: ["success", "failure"]
          description: Execution result status
          example: "success"
        labels:
          type: object
          additionalProperties:
            type: string
          description: Additional labels for this execution
          example:
            env: "prod"
            backup_type: "full"
        duration:
          type: integer
          description: Execution duration in seconds
          example: 120
        output:
          type: string
          description: Optional execution output or error message
          example: "Backup completed successfully. 1.2GB transferred."
        timestamp:
          type: string
          format: date-time
          description: Execution timestamp (defaults to current time if not provided)
          example: "2025-10-30T19:56:00Z"
      required:
        - job_name
        - host
        - status
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Job result submitted successfully"
        timestamp:
          type: string
          format: date-time
          example: "2025-10-30T19:56:00Z"
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Job name and host are required"
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: "2025-10-30T19:56:00Z"
      required:
        - error
        - timestamp
...
  responses:
...
    BadRequestError:
      description: Bad request - invalid input data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Job name and host are required"
            timestamp: "2025-10-30T19:56:00Z"
    UnauthorizedError:
      description: Unauthorized - missing or invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "missing or invalid API key"
            timestamp: "2025-10-30T19:56:00Z"
...
    ConflictError:
      description: Conflict - resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "job already exists"
            timestamp: "2025-10-30T19:56:00Z"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "failed to create job"
            timestamp: "2025-10-30T19:56:00Z"
...