# pkg/api/server.go
package api

import (
	"encoding/json"
	"errors"
	"fmt"
	"net/http"
	"os"
	"path/filepath"
	"strings"
	"time"

	httpSwagger "github.com/swaggo/http-swagger/v2"

	"github.com/jaepetto/cron-exporter/pkg/config"
	"github.com/jaepetto/cron-exporter/pkg/dashboard"
	"github.com/jaepetto/cron-exporter/pkg/jobservice"
	"github.com/jaepetto/cron-exporter/pkg/metrics"
	"github.com/jaepetto/cron-exporter/pkg/model"
	"github.com/sirupsen/logrus"
)

// Server represents the HTTP API server
type Server struct {
	config         *config.Config
	jobStore       *model.JobStore
	jobService     *jobservice.Service
	jobResultStore *model.JobResultStore
	metrics        *metrics.Collector
	dashboard      *dashboard.Dashboard
}

// NewServer creates a new API server instance
func NewServer(cfg *config.Config, jobStore *model.JobStore, jobResultStore *model.JobResultStore, metricsCollector *metrics.Collector) *Server {
	server := &Server{
		config:         cfg,
		jobStore:       jobStore,
		jobService:     jobservice.New(jobStore),
		jobResultStore: jobResultStore,
		metrics:        metricsCollector,
	}

	// Initialize dashboard if enabled
	if cfg.Dashboard.Enabled {
		server.dashboard = dashboard.New(
			&cfg.Dashboard,
			jobStore,
			cfg.Security.AdminAPIKeys,
			logrus.StandardLogger(),
		)
	}

	return server
}

// Handler returns the HTTP handler for the server
func (s *Server) Handler() http.Handler {
	mux := http.NewServeMux()

	// API routes
	mux.HandleFunc("/api/job", s.withAuth(s.handleJob))
	mux.HandleFunc("/api/job/", s.withAuth(s.handleJobByID))
	mux.HandleFunc("/api/job-result", s.withJobAuth(s.handleJobResult))

	// Metrics endpoint
	mux.HandleFunc(s.config.Metrics.Path, s.handleMetrics)

	// Health check
	mux.HandleFunc("/health", s.handleHealth)

	// Swagger UI and OpenAPI spec
	mux.Handle("/swagger/", httpSwagger.Handler(
		httpSwagger.URL("/api/openapi.yaml"), // The URL pointing to the OpenAPI spec
		httpSwagger.DeepLinking(true),
		httpSwagger.DocExpansion("list"),
		httpSwagger.DomID("swagger-ui"),
	))
	mux.HandleFunc("/api/openapi.yaml", s.handleOpenAPISpec)

	// Mount dashboard if enabled
	if s.dashboard != nil && s.dashboard.IsEnabled() {
		mux.Handle(s.config.Dashboard.Path, http.RedirectHandler(s.config.Dashboard.Path+"/", http.StatusPermanentRedirect))
		mux.Handle(s.config.Dashboard.Path+"/", http.StripPrefix(s.config.Dashboard.Path, s.dashboard.Router()))
	}

	// Add request logging middleware
	return s.withLogging(mux)
}

// withAuth provides authentication middleware for admin operations
func (s *Server) withAuth(handler http.HandlerFunc) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		// Skip auth in development mode
		if s.config.Database.Path == "/tmp/cronmetrics_dev.db" {
			handler(w, r)
			return
		}

		// Get API key from header
		apiKey := s.extractAPIKey(r)
		if apiKey == "" {
			s.writeErrorResponse(w, http.StatusUnauthorized, "missing or invalid API key")
			return
		}

		// Check if token is valid admin key
		if !s.isValidAdminAPIKey(apiKey) {
			s.writeErrorResponse(w, http.StatusUnauthorized, "admin access required")
			return
		}

		// Add auth info to request context
		r.Header.Set("X-Auth-Level", "admin")
		handler(w, r)
	}
}

// withJobAuth provides authentication middleware for job result submissions
func (s *Server) withJobAuth(handler http.HandlerFunc) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		// Skip auth in development mode
		if s.config.Database.Path == "/tmp/cronmetrics_dev.db" {
			handler(w, r)
			return
		}

		// Get API key from header
		apiKey := s.extractAPIKey(r)
		if apiKey == "" {
			s.writeErrorResponse(w, http.StatusUnauthorized, "missing or invalid API key")
			return
		}

		// Validate API key by looking up the associated job
		job, err := s.jobStore.GetJobByApiKey(apiKey)
		if err != nil {
			s.writeErrorResponse(w, http.StatusUnauthorized, "invalid API key")
			return
		}

		// Add job info to request context for validation
		r.Header.Set("X-Auth-Job-Name", job.Name)
		r.Header.Set("X-Auth-Job-Host", job.Host)
		r.Header.Set("X-Auth-Level", "job")

		handler(w, r)
	}
}

// extractAPIKey extracts API key from various header formats
func (s *Server) extractAPIKey(r *http.Request) string {
	// Try X-API-Key header first (preferred for job submissions)
	if apiKey := r.Header.Get("X-API-Key"); apiKey != "" {
		return apiKey
	}

	// Fall back to Authorization Bearer format (for admin operations)
	authHeader := r.Header.Get("Authorization")
	if authHeader == "" {
		return ""
	}

	// Extract Bearer token
	token := strings.TrimPrefix(authHeader, "Bearer ")
	if token == authHeader {
		return ""
	}

	return token
}

// withLogging provides request logging middleware
func (s *Server) withLogging(handler http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		start := time.Now()

		// Create a wrapped response writer to capture status code
		wrapped := &responseWriter{ResponseWriter: w, statusCode: http.StatusOK}

		handler.ServeHTTP(wrapped, r)

		duration := time.Since(start)

		logrus.WithFields(logrus.Fields{
			"method":         r.Method,
			"path":           r.URL.Path,
			"status":         wrapped.statusCode,
			"duration_ms":    duration.Milliseconds(),
			"remote_addr":    r.RemoteAddr,
			"user_agent":     r.UserAgent(),
			"content_length": r.ContentLength,
		}).Info("http request")
	})
}

// responseWriter wraps http.ResponseWriter to capture status code
type responseWriter struct {
	http.ResponseWriter
	statusCode int
}

func (rw *responseWriter) Flush() {
	if flusher, ok := rw.ResponseWriter.(http.Flusher); ok {
		flusher.Flush()
	}
}

func (rw *responseWriter) WriteHeader(code int) {
	rw.statusCode = code
	rw.ResponseWriter.WriteHeader(code)
}

// handleJob handles job CRUD operations
func (s *Server) handleJob(w http.ResponseWriter, r *http.Request) {
	switch r.Method {
	case http.MethodPost:
		s.handleCreateJob(w, r)
	case http.MethodGet:
		s.handleListJobs(w, r)
	default:
		s.writeErrorResponse(w, http.StatusMethodNotAllowed, "method not allowed")
	}
}

// handleJobByID handles operations on specific jobs using job ID
func (s *Server) handleJobByID(w http.ResponseWriter, r *http.Request) {
	// Extract job ID from path
	path := strings.TrimPrefix(r.URL.Path, "/api/job/")

	if path == "" {
		s.writeErrorResponse(w, http.StatusBadRequest, "invalid job path format (expected /api/job/{id})")
		return
	}

	// Parse job ID
	jobID := 0
	if _, err := fmt.Sscanf(path, "%d", &jobID); err != nil {
		s.writeErrorResponse(w, http.StatusBadRequest, "invalid job ID format (must be a number)")
		return
	}

	switch r.Method {
	case http.MethodGet:
		s.handleGetJobByID(w, r, jobID)
	case http.MethodPut:
		s.handleUpdateJobByID(w, r, jobID)
	case http.MethodDelete:
		s.handleDeleteJobByID(w, r, jobID)
	default:
		s.writeErrorResponse(w, http.StatusMethodNotAllowed, "method not allowed")
	}
}

// handleCreateJob creates a new job
func (s *Server) handleCreateJob(w http.ResponseWriter, r *http.Request) {
	// Only admin can create jobs
	if r.Header.Get("X-Auth-Level") != "admin" {
		s.writeErrorResponse(w, http.StatusForbidden, "admin access required")
		return
	}

	var input jobservice.CreateJobInput
	if err := json.NewDecoder(r.Body).Decode(&input); err != nil {
		s.writeErrorResponse(w, http.StatusBadRequest, fmt.Sprintf("invalid JSON: %v", err))
		return
	}

	job, err := s.jobService.Create(input)
	if err != nil {
		s.writeJobServiceError(w, err)
		return
	}

	s.writeJSONResponse(w, http.StatusCreated, job)
}

// handleListJobs lists all jobs with optional filtering
func (s *Server) handleListJobs(w http.ResponseWriter, r *http.Request) {
	// Parse label filters from query parameters
	labelFilters := make(map[string]string)
	for key, values := range r.URL.Query() {
		if strings.HasPrefix(key, "label.") {
			labelKey := strings.TrimPrefix(key, "label.")
			if len(values) > 0 {
				labelFilters[labelKey] = values[0]
			}
		}
	}

	jobs, err := s.jobStore.ListJobs(labelFilters)
	if err != nil {
		s.writeErrorResponse(w, http.StatusInternalServerError, fmt.Sprintf("failed to list jobs: %v", err))
		return
	}

	s.writeJSONResponse(w, http.StatusOK, jobs)
}

// handleGetJobByID retrieves a specific job by ID
func (s *Server) handleGetJobByID(w http.ResponseWriter, r *http.Request, jobID int) {
	job, err := s.jobService.Get(jobID)
	if err != nil {
		s.writeJobServiceError(w, err)
		return
	}

	s.writeJSONResponse(w, http.StatusOK, job)
}

// handleGetJob retrieves a specific job (kept for backward compatibility)
func (s *Server) handleGetJob(w http.ResponseWriter, r *http.Request, jobName, jobHost string) {
	job, err := s.jobStore.GetJob(jobName, jobHost)
	if err != nil {
		if strings.Contains(err.Error(), "not found") {
			s.writeErrorResponse(w, http.StatusNotFound, "job not found")
			return
		}
		s.writeErrorResponse(w, http.StatusInternalServerError, fmt.Sprintf("failed to get job: %v", err))
		return
	}

	s.writeJSONResponse(w, http.StatusOK, job)
}

// handleUpdateJobByID updates a job by ID
func (s *Server) handleUpdateJobByID(w http.ResponseWriter, r *http.Request, jobID int) {
	// Only admin can update jobs
	if r.Header.Get("X-Auth-Level") != "admin" {
		s.writeErrorResponse(w, http.StatusForbidden, "admin access required")
		return
	}

	var input jobservice.UpdateJobInput
	if err := json.NewDecoder(r.Body).Decode(&input); err != nil {
		s.writeErrorResponse(w, http.StatusBadRequest, fmt.Sprintf("invalid JSON: %v", err))
		return
	}

	job, err := s.jobService.Update(jobID, input)
	if err != nil {
		s.writeJobServiceError(w, err)
		return
	}

	s.writeJSONResponse(w, http.StatusOK, job)
}

// handleUpdateJob updates a job (kept for backward compatibility)
func (s *Server) handleUpdateJob(w http.ResponseWriter, r *http.Request, jobName, jobHost string) {
	// Only admin can update jobs
	if r.Header.Get("X-Auth-Level") != "admin" {
		s.writeErrorResponse(w, http.StatusForbidden, "admin access required")
		return
	}

	// Get existing job
	existingJob, err := s.jobStore.GetJob(jobName, jobHost)
	if err != nil {
		if strings.Contains(err.Error(), "not found") {
			s.writeErrorResponse(w, http.StatusNotFound, "job not found")
			return
		}
		s.writeErrorResponse(w, http.StatusInternalServerError, fmt.Sprintf("failed to get job: %v", err))
		return
	}

	var updateData model.Job
	if err := json.NewDecoder(r.Body).Decode(&updateData); err != nil {
		s.writeErrorResponse(w, http.StatusBadRequest, fmt.Sprintf("invalid JSON: %v", err))
		return
	}

	// Update only provided fields
	if updateData.ApiKey != "" {
		existingJob.ApiKey = updateData.ApiKey
	}
	if updateData.AutomaticFailureThreshold > 0 {
		existingJob.AutomaticFailureThreshold = updateData.AutomaticFailureThreshold
	}
	if updateData.Labels != nil {
		existingJob.Labels = updateData.Labels
	}
	if updateData.Status != "" {
		existingJob.Status = updateData.Status
	}

	if err := s.jobStore.UpdateJob(existingJob); err != nil {
		s.writeErrorResponse(w, http.StatusInternalServerError, fmt.Sprintf("failed to update job: %v", err))
		return
	}

	s.writeJSONResponse(w, http.StatusOK, existingJob)
}

// handleDeleteJobByID deletes a job by ID
func (s *Server) handleDeleteJobByID(w http.ResponseWriter, r *http.Request, jobID int) {
	// Only admin can delete jobs
	if r.Header.Get("X-Auth-Level") != "admin" {
		s.writeErrorResponse(w, http.StatusForbidden, "admin access required")
		return
	}

	if _, err := s.jobService.Delete(jobID); err != nil {
		s.writeJobServiceError(w, err)
		return
	}

	w.WriteHeader(http.StatusNoContent)
}

// handleDeleteJob deletes a job (kept for backward compatibility)
func (s *Server) handleDeleteJob(w http.ResponseWriter, r *http.Request, jobName, jobHost string) {
	// Only admin can delete jobs
	if r.Header.Get("X-Auth-Level") != "admin" {
		s.writeErrorResponse(w, http.StatusForbidden, "admin access required")
		return
	}

	if err := s.jobStore.DeleteJob(jobName, jobHost); err != nil {
		if strings.Contains(err.Error(), "not found") {
			s.writeErrorResponse(w, http.StatusNotFound, "job not found")
			return
		}
		s.writeErrorResponse(w, http.StatusInternalServerError, fmt.Sprintf("failed to delete job: %v", err))
		return
	}

	w.WriteHeader(http.StatusNoContent)
}

// handleJobResult handles job result submissions
func (s *Server) handleJobResult(w http.ResponseWriter, r *http.Request) {
	if r.Method != http.MethodPost {
		s.writeErrorResponse(w, http.StatusMethodNotAllowed, "method not allowed")
		return
	}

	var result model.JobResult
	if err := json.NewDecoder(r.Body).Decode(&result); err != nil {
		s.writeErrorResponse(w, http.StatusBadRequest, fmt.Sprintf("invalid JSON: %v", err))
		return
	}

	// Validate required fields
	if result.JobName == "" || result.Host == "" || result.Status == "" {
		s.writeErrorResponse(w, http.StatusBadRequest, "job_name, host, and status are required")
		return
	}

	// Validate status
	if result.Status != "success" && result.Status != "failure" {
		s.writeErrorResponse(w, http.StatusBadRequest, "status must be 'success' or 'failure'")
		return
	}

	// In non-dev mode, validate that the job result matches the authenticated job
	if s.config.Database.Path != "/tmp/cronmetrics_dev.db" {
		authJobName := r.Header.Get("X-Auth-Job-Name")
		authJobHost := r.Header.Get("X-Auth-Job-Host")

		if result.JobName != authJobName || result.Host != authJobHost {
			s.writeErrorResponse(w, http.StatusForbidden, "job result does not match authenticated job")
			return
		}
	}

	// Set timestamp if not provided
	if result.Timestamp.IsZero() {
		result.Timestamp = time.Now().UTC()
	}

	// Store the job result
	if err := s.jobResultStore.CreateJobResult(&result); err != nil {
		s.writeErrorResponse(w, http.StatusInternalServerError, fmt.Sprintf("failed to store job result: %v", err))
		return
	}

	// Update job's last reported timestamp
	if err := s.jobStore.UpdateJobLastReported(result.JobName, result.Host, result.Timestamp); err != nil {
		// Log error but don't fail the request
		logrus.WithError(err).WithFields(logrus.Fields{
			"job_name": result.JobName,
			"host":     result.Host,
		}).Warn("failed to update job last reported timestamp")
	}

	// Broadcast job status change to dashboard clients if dashboard is enabled
	if s.dashboard != nil && s.dashboard.IsEnabled() {
		if broadcaster := s.dashboard.GetBroadcaster(); broadcaster != nil {
			// Get the updated job to broadcast current status
			if job, err := s.jobStore.GetJob(result.JobName, result.Host); err == nil {
				// Determine if job is in failure state
				isFailure := result.Status == "failure"

				// Also check automatic failure threshold
				if !isFailure && job.AutomaticFailureThreshold > 0 {
					timeSinceLastReport := time.Since(job.LastReportedAt)
					if timeSinceLastReport > time.Duration(job.AutomaticFailureThreshold)*time.Second {
						isFailure = true
					}
				}

				broadcaster.BroadcastJobStatusChange(job, isFailure)
			}
		}
	}

	s.writeJSONResponse(w, http.StatusCreated, map[string]string{
		"status": "recorded",
		"job":    fmt.Sprintf("%s@%s", result.JobName, result.Host),
	})
}

// handleMetrics serves Prometheus metrics
func (s *Server) handleMetrics(w http.ResponseWriter, r *http.Request) {
	if r.Method != http.MethodGet {
		s.writeErrorResponse(w, http.StatusMethodNotAllowed, "method not allowed")
		return
	}

	metrics, err := s.metrics.Gather()
	if err != nil {
		s.writeErrorResponse(w, http.StatusInternalServerError, fmt.Sprintf("failed to gather metrics: %v", err))
		return
	}

	w.Header().Set("Content-Type", "text/plain; version=0.0.4; charset=utf-8")
	w.WriteHeader(http.StatusOK)
	if _, err := w.Write([]byte(metrics)); err != nil {
		logrus.WithError(err).Error("Failed to write metrics response")
	}
}

// handleHealth handles health check requests
func (s *Server) handleHealth(w http.ResponseWriter, r *http.Request) {
	if r.Method != http.MethodGet {
		s.writeErrorResponse(w, http.StatusMethodNotAllowed, "method not allowed")
		return
	}

	health := map[string]interface{}{
		"status":    "healthy",
		"timestamp": time.Now().UTC().Format(time.RFC3339),
		"version":   "0.1.0",
	}

	s.writeJSONResponse(w, http.StatusOK, health)
}

// handleOpenAPISpec serves the OpenAPI specification file
func (s *Server) handleOpenAPISpec(w http.ResponseWriter, r *http.Request) {
	if r.Method != http.MethodGet {
		s.writeErrorResponse(w, http.StatusMethodNotAllowed, "method not allowed")
		return
	}

	// Find the OpenAPI spec file relative to the binary
	specPath := "docs/openapi.yaml"

	// Try to find the spec file in multiple locations
	possiblePaths := []string{
		specPath,                            // Relative to current working directory
		filepath.Join(".", specPath),        // Explicit relative path
		filepath.Join("..", specPath),       // One level up (in case running from bin/)
		filepath.Join("..", "..", specPath), // Two levels up
	}

	var content []byte
	var err error

	for _, path := range possiblePaths {
		// Validate path to prevent directory traversal attacks
		cleanPath := filepath.Clean(path)

		// Must end with openapi.yaml to prevent reading arbitrary files
		if !strings.HasSuffix(cleanPath, "openapi.yaml") {
			continue
		}

		// Convert to absolute path and validate it ends with docs/openapi.yaml
		absPath, pathErr := filepath.Abs(cleanPath)
		if pathErr != nil {
			continue
		}

		// Ensure the absolute path contains the expected docs/openapi.yaml structure
		if !strings.HasSuffix(absPath, "/docs/openapi.yaml") && !strings.HasSuffix(absPath, "\\docs\\openapi.yaml") {
			continue
		}

		content, err = os.ReadFile(cleanPath) // #nosec G304 - path validated above
		if err == nil {
			break
		}
	}

	if err != nil {
		logrus.WithError(err).Errorf("Failed to read OpenAPI spec from any of these paths: %v", possiblePaths)
		s.writeErrorResponse(w, http.StatusInternalServerError, "OpenAPI specification not found")
		return
	}

	w.Header().Set("Content-Type", "application/yaml")
	w.Header().Set("Cache-Control", "public, max-age=3600") // Cache for 1 hour
	w.WriteHeader(http.StatusOK)
	if _, err := w.Write(content); err != nil {
		logrus.WithError(err).Error("Failed to write OpenAPI spec response")
	}
}

// isValidAdminAPIKey checks if the provided token is a valid admin API key
func (s *Server) isValidAdminAPIKey(token string) bool {
	for _, key := range s.config.Security.AdminAPIKeys {
		if key == token {
			return true
		}
	}
	return false
}

// writeJSONResponse writes a JSON response
func (s *Server) writeJSONResponse(w http.ResponseWriter, statusCode int, data interface{}) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(statusCode)

	if err := json.NewEncoder(w).Encode(data); err != nil {
		logrus.WithError(err).Error("failed to encode JSON response")
	}
}

// writeErrorResponse writes an error response
func (s *Server) writeErrorResponse(w http.ResponseWriter, statusCode int, message string) {
	errorResponse := map[string]string{
		"error":     message,
		"timestamp": time.Now().UTC().Format(time.RFC3339),
	}

	s.writeJSONResponse(w, statusCode, errorResponse)
}

func (s *Server) writeJobServiceError(w http.ResponseWriter, err error) {
	var serviceError *jobservice.Error
	if !errors.As(err, &serviceError) {
		s.writeErrorResponse(w, http.StatusInternalServerError, "internal server error")
		return
	}

	statusCode := http.StatusInternalServerError
	switch serviceError.Kind {
	case jobservice.ErrorValidation:
		statusCode = http.StatusBadRequest
	case jobservice.ErrorNotFound:
		statusCode = http.StatusNotFound
	case jobservice.ErrorConflict:
		statusCode = http.StatusConflict
	}

	response := map[string]interface{}{
		"error":     serviceError.Message,
		"code":      serviceError.Kind,
		"timestamp": time.Now().UTC().Format(time.RFC3339),
	}
	if len(serviceError.Fields) > 0 {
		response["fields"] = serviceError.Fields
	}
	s.writeJSONResponse(w, statusCode, response)
}